Beta: lodg: MQTT for HASS complete! - Removed JSON lib as not needed (plain string formatting is enough) - Added 'measurement unit' configuration to devices - Metric values are now floats - Reworked lots of data structures

+++ lodg/src/gatherers/headers/gath_unix.h
@@ -5,7 +5,7 @@
 
 #include <stdint.h>
 
-uint8_t get_la1();
+float get_la1();
 
 #endif
